Fix impl issue for user utils
